Start importing data from a CSV file to the database in a job

+++ src/test/java/com/ecom/shopease/SettingGame.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @AllArgsConstructor
 @Data
 public class SettingGame {
-    private int maxPlayers;
+    private double maxPlayers;
     private String mode;
 }
